(SIMP-3430) Create Centralize spec_helpers.rb
- created baseline for spec_helper.rb
  to do basc RSPEC configuration for SIMP module unit tests.
- added some basic methods that were also in spec_helpers
- added compliance markup methods for use in compliance
  markup unit tests.
- note that you need compliance_markup in the fixtures file for
  these helpers to work.

SIMP-3430 #comment added simp-spec-helpers
